% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSExact_v1.R
\name{GSExact_basic}
\alias{GSExact_basic}
\title{GSExact function
This is Testing. It's my very first time uploading package onto Github}
\usage{
GSExact_basic(input_data, J, n, burn, iter)
}
\description{
GSExact function
This is Testing. It's my very first time uploading package onto Github
}
